Move LineChart demos to LineChartDemo namespace
